Streamline
I decided “kindly” was inessential to the definition because if you view something as important, you’ll naturally treat it sympathetically, helpfully, gently, and patiently. Sure, it’s weird to say that because you see the Civil War as “important”—an incontrovertible fact for Americans—and find it weird to say you love it,  but the minute you begin “treating“ it as important, I think we could all agree say you “love“ it.